Refactor to module below src & apply requisite touchups
